+++ allauth/socialaccount/providers/discogs/urls.py
@@ -1,0 +1,5 @@
+from allauth.socialaccount.providers.discogs.provider import DiscogsProvider
+from allauth.socialaccount.providers.oauth.urls import default_urlpatterns
+
+
+urlpatterns = default_urlpatterns(DiscogsProvider)