+++ tests/projects/common/settings.py
@@ -26,6 +26,7 @@ INSTALLED_SOCIALACCOUNT_APPS: Tuple[str, ...] = (
     "allauth.socialaccount.providers.daum",
     "allauth.socialaccount.providers.digitalocean",
     "allauth.socialaccount.providers.dingtalk",
+    "allauth.socialaccount.providers.discogs",
     "allauth.socialaccount.providers.discord",
     "allauth.socialaccount.providers.disqus",
     "allauth.socialaccount.providers.douban",